Move HPO tab code into separate controller, fix bug with property value removal. Add optional resources for Mondo support
