+++ hpowbgui/src/main/resources/resource_bundle/ResourceBundle.properties
@@ -1,3 +1,4 @@
 status.all.set=All resources set!
 status.download.annotations=Please download HPO annotations file (Edit | Download HPO annotations) 
 status.download.hpo=Please download HPO obo file (Edit | Download HPO) or set path to local one (Edit | Import local copy...)
+status.download.mondo=Please download MONDO obo file (Edit | Download Mondo)
